refactor default filter logic
